moved all functions from main.rs to lib.rs and replaced a nasty clone with an iterator.
